+++ parsers/lamoda_parser.py
@@ -1,9 +1,13 @@
+import logging
+from decimal import Decimal
+
 import aiohttp
 from bs4 import BeautifulSoup
+from bson import Decimal128
 from fastapi import HTTPException
 
 from services.lamoda_db import LamodaServiceDatabase
-
+logger = logging.getLogger(__name__)
 lamoda_mongo = LamodaServiceDatabase()
 
 lamoda_url = "https://www.lamoda.by"
@@ -42,3 +46,80 @@ async def get_urls_categories(session):
             list_urls.append(lamoda_url + category["href"])
 
     return list_urls
+
+
+async def fetch_item(url, session):
+    async with session.get(url, headers=None) as response:
+        soup = BeautifulSoup(await response.text(), "lxml")
+
+        color = str()
+        product_info = soup.findAll("script")[6].text.strip()
+        color_index = soup.findAll("script")[6].text.strip().find("Цвет")
+        if color_index == -1:
+            color = "Не указан"
+        else:
+            color_index += 15
+
+            while product_info[color_index] != '"':
+                color += product_info[color_index]
+                color_index += 1
+
+        category = soup.findAll("a", class_="x-link x-link__secondaryLabel")[
+            2
+        ]
+        brand = soup.find(
+            "span", class_="x-premium-product-title__brand-name"
+        ).text.strip()
+        name = soup.find(
+            "div", class_="x-premium-product-title__model-name"
+        ).text.strip()
+        if soup.find("span", class_="x-premium-product-prices__price") is None:
+            return None
+        price_split = soup.find(
+            "span", class_="x-premium-product-prices__price"
+        ).text.split(" ")
+        price = (
+            "".join(price_split[:-1])
+            if price_split[-1] == "р."
+            else "".join(price_split)
+        )
+        info_value = soup.findAll(
+            "span", class_="x-premium-product-description-attribute__value"
+        )
+        info_name = soup.findAll(
+            "span", class_="x-premium-product-description-attribute__name"
+        )
+        info_result = {}
+
+        for i, d in zip(info_value, info_name):
+            info_result[d.text] = i.text
+        article = info_result["Артикул"]
+
+        item = ItemModel(
+            name=name,
+            article=article,
+            category=category.text.strip(),
+            price=Decimal128(Decimal(price)),
+            brand=brand,
+            color=color.capitalize(),
+        )
+        logger.info(f"Item info: {item}")
+
+        return item
+
+
+async def fetch_category_items(category_url, client):
+    items = []
+    async with client.get(category_url, headers=None) as response:
+        soup = BeautifulSoup(await response.text(), "lxml")
+
+        items_url = soup.findAll(
+            "a", class_="x-product-card__link x-product-card__hit-area"
+        )
+
+        for item_url in items_url:
+            fetch_items = await fetch_item(lamoda_url + item_url["href"], client)
+            if fetch_items is not None:
+                items.append(fetch_items)
+
+        return items
